@@ -5,30 +5,30 @@
   "license": "MIT",
   "private": false,
   "dependencies": {
-    "@types/react": "^17.0.25",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "bootstrap": "5.1.1",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "bootstrap": "5.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-ga": "^3.3.0",
     "react-helmet-async": "^1.1.2",
-    "react-router-dom": "^5.3.0",
-    "web-vitals": "^2.1.0"
+    "react-router-dom": "^6.0.2",
+    "web-vitals": "^2.1.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.1",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.10.2",
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.10",
     "@types/react-test-renderer": "^17.0.1",
     "gh-pages": "^3.2.3",
     "postcss-custom-properties": "^10.0.0",
     "react-scripts": "^4.0.3",
     "react-test-renderer": "^17.0.2",
-    "sass": "^1.42.1",
-    "typescript": "^4.4.3"
+    "sass": "^1.43.5",
+    "typescript": "^4.5.2"
   },
   "scripts": {
     "predeploy": "npm run build",
